+++ src/Altinn.Correspondence.Persistence/Repositories/CorrespondenceRepository.cs
@@ -389,6 +389,10 @@ namespace Altinn.Correspondence.Persistence.Repositories
             {
                 return 0;
             }
+            if (entities.Count > 1000) // Safety margin
+            {
+                throw new ArgumentException($"Too many correspondences to delete. Total correspondences in requested hard delete operation: {entities.Count}");
+            }
 
             _context.Correspondences.RemoveRange(entities);
             await _context.SaveChangesAsync(cancellationToken);
